CA1707: Identifiers should not contain underscores (https://docs.microsoft.com/en-us/visualstudio/code-quality/ca1707-identifiers-should-not-contain-underscores)
It should be noted that this warning only appeard on property identifiers, not field or variable identifiers. This appears to be a bug with the FxCop implementation given that the documentation says it checks members and a field is a member. We should do a deeper clean here. This commit eliminates the warning message as the smallest possible change.