CS0168: Unused variables
This warning is generated by the compiler, not the code analysis tool. In each case of this commit the unused variable was an exception captured by the logic and then swallowed. Instead of swallowing these errors we should log and possibly re-throw the errors. That change is outside the scope of this effort.